Split node_sbg_translator.cpp into a hpp and a cpp.
